refactor(editor): move point/simplify helpers to connection-waypoints
