+++ apps/frontend/src/view/components/editor-components/connection-routing/shared.test.ts
@@ -3,12 +3,10 @@ import {
     compareRouteDefects,
     countObstacleHits,
     faceMidpoint,
-    flattenPoints,
     rectangleOf,
     routeLength,
     segmentHitsRectangle,
     shrinkRectangle,
-    simplifyPolyline,
     stepDirection,
 } from "./shared.ts";
 import { createSystemComponent, createPointOfAttack } from "#test-utils/builders.ts";
@@ -42,48 +40,6 @@ describe("faceMidpoint", () => {
         expect(faceMidpoint(component, AnchorOrientation.right)).toEqual({ x: 130, y: 140 });
         expect(faceMidpoint(component, AnchorOrientation.top)).toEqual({ x: 90, y: 100 });
         expect(faceMidpoint(component, AnchorOrientation.bottom)).toEqual({ x: 90, y: 180 });
-    });
-});
-
-describe("simplifyPolyline", () => {
-    it("collapses a straight run to its endpoints", () => {
-        const simplified = simplifyPolyline([
-            { x: 0, y: 0 },
-            { x: 10, y: 0 },
-            { x: 20, y: 0 },
-        ]);
-
-        expect(simplified).toEqual([
-            { x: 0, y: 0 },
-            { x: 20, y: 0 },
-        ]);
-    });
-
-    it("keeps the corner of an L-shaped path", () => {
-        const simplified = simplifyPolyline([
-            { x: 0, y: 0 },
-            { x: 10, y: 0 },
-            { x: 10, y: 10 },
-        ]);
-
-        expect(simplified).toEqual([
-            { x: 0, y: 0 },
-            { x: 10, y: 0 },
-            { x: 10, y: 10 },
-        ]);
-    });
-
-    it("drops consecutive duplicate points", () => {
-        const simplified = simplifyPolyline([
-            { x: 0, y: 0 },
-            { x: 0, y: 0 },
-            { x: 10, y: 0 },
-        ]);
-
-        expect(simplified).toEqual([
-            { x: 0, y: 0 },
-            { x: 10, y: 0 },
-        ]);
     });
 });
 
@@ -183,18 +139,6 @@ describe("stepDirection", () => {
         expect(stepDirection({ x: 0, y: 0 }, { x: 5, y: 0 })).toEqual({ x: 1, y: 0 });
         expect(stepDirection({ x: 0, y: 0 }, { x: 0, y: -5 })).toEqual({ x: 0, y: -1 });
         expect(stepDirection({ x: 5, y: 5 }, { x: 5, y: 5 })).toEqual({ x: 0, y: 0 });
-    });
-});
-
-describe("flattenPoints", () => {
-    it("turns points into the flat array Konva's Line expects", () => {
-        expect(
-            flattenPoints([
-                { x: 0, y: 0 },
-                { x: 10, y: 20 },
-                { x: 10, y: 50 },
-            ])
-        ).toEqual([0, 0, 10, 20, 10, 50]);
     });
 });
